requirements.txt
tensorflow<2.11
numpy==1.26.4
pillow==11.3.0
natsort==8.4.0